Add and update various constants
Expanded menu text to add a fifth option.
Added a new string to print when packing is attempted prior to boxes
being loaded.
Added paths for metrics output files.
Added header for Option 1 output files.
Added a new string to be displayed when prompting for the
identifier of the last output file.