Modify write_input_box_func
Generating a color for a box has been moved 
out of the Option1-specific block. 
There was a bug where boxes of the same type 
had different dimensions for attributes that 
should be identical. Code generating these 
values have been moved out of the inner for 
loop to correct this behavior.